Agregar endpoints para Producto

+++ Models/Dtos/ProductoDto.cs
@@ -13,8 +13,8 @@ namespace Models.Dtos
 
         public string Marca { get; set; }
 
-        public double Precio { get; set; }
+        public decimal Precio { get; set; }
 
-        public double Costo { get; set; }
+        public decimal Costo { get; set; }
     }
 }
